pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Heta"
version = "0.1.0"
description = "Heterogeneous Data Management Platform"
readme = "README.md"
requires-python = ">=3.10,<3.12"
license = {text = "MIT"}
authors = [
    {name = "HetaTeam"}
]

dependencies = [
    "numpy==1.26.4",
    "pandas==2.3.3",
    "scikit-learn==1.7.2",
    "fastapi==0.133.1",
    "uvicorn==0.41.0",
    "pydantic==2.12.5",
    "httpx==0.28.1",
    # httpx[socks] extra (socksio) is required when the host machine uses a SOCKS
    # proxy (e.g. ALL_PROXY=socks://...). Without it, MemoryVG (LightRAG/httpx)
    # and any other httpx client will raise "Unknown scheme for proxy URL" on init.
    # Install manually: pip install "httpx[socks]"
    # Not listed as a hard dependency because it is only needed in proxy environments.
    "requests==2.32.5",
    "torch==2.9.1",
    "torchvision==0.24.1",
    "openai==1.109.1",
    "langchain==0.3.27",
    "langchain-core==0.3.83",
    "langchain-ollama==0.3.10",
    "qwen-agent==0.0.31",
    "ollama==0.6.1",
    "tiktoken==0.12.0",
    "transformers==4.57.6",
    "pymilvus==2.6.9",
    "psycopg2-binary==2.9.11",
    "SQLAlchemy==2.0.47",
    "mineru==2.2.2",
    "onnxruntime==1.23.2",
    "ftfy==6.3.1",
    "docling==2.75.0",
    "beautifulsoup4==4.14.3",
    "lxml==6.0.2",
    "Pillow==12.1.1",
    "easyocr==1.7.2",
    "rarfile==4.2",
    "py7zr==1.1.0",
    "python-dotenv==1.2.1",
    "PyYAML==6.0.3",
    "ruamel.yaml==0.19.1",
    "loguru==0.7.3",
    "tqdm==4.67.3",
    "tabulate==0.9.0",
    "streamlit==1.54.0",
    "rich==14.3.3",
    "playwright==1.58.0",
    "imap-tools==1.11.1",
    "click==8.3.1",
    "colorama==0.4.6",
    "colorlog==6.10.1",
    "doclayout_yolo==0.0.4",
    "ultralytics==8.4.8",
    "python-multipart==0.0.22",
    "sentence-transformers==5.2.0",
    "jieba==0.42.1",
    "zhconv==1.4.3",
    # hetamem dependencies
    "pytz==2025.2",
    "posthog==7.9.4",
    "qdrant-client==1.17.0",
    "langchain-neo4j==0.3.0",
    "rank-bm25==0.2.2",
    "json-repair==0.58.0",
    "pipmaster==1.1.1",
    "nano-vectordb==0.0.4.3",
    "networkx==3.4.2",
    "scipy==1.15.3",
    "python-louvain==0.16",
    "mcp==1.26.0",
]

[project.optional-dependencies]
hetamem-viz = [
    "imgui-bundle>=1.0.0",
    "moderngl>=5.8.0",
    "pyglm>=2.7.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.1.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]
gpu = [
    "torch[cuda]>=2.0.0",
]
docs = [
    "sphinx>=5.0",
    "sphinx-rtd-theme>=1.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["hetamem", "hetamem.*", "common", "common.*", "hetadb", "hetadb.*", "hetagen", "hetagen.*"]


[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
